Adding linting for style and content
