+++ client/src/App.js
@@ -1,11 +1,11 @@
-import React from "react";
-import { BrowserRouter, Switch, Route } from "react-router-dom";
-import LogoBar from "./components/LogoBar";
+import React from 'react';
+import { BrowserRouter, Switch, Route } from 'react-router-dom';
+import LogoBar from './components/LogoBar';
 import FrontPage from './views/FrontPage';
-import EventDetailPage from "./views/EventDetailPage";
-import CreateEvent from "./components/Events/CreateEvent";
-import EventSubmitted from "./components/Events/EventSubmitted";
-import Banner from "./components/Banner";
+import EventDetailPage from './views/EventDetailPage';
+import CreateEvent from './components/Events/CreateEvent';
+import EventSubmitted from './components/Events/EventSubmitted';
+import Banner from './components/Banner';
 
 function App() {
   return (
@@ -24,6 +24,8 @@ function App() {
         <Route path="/event/:id">
           <EventDetailPage />
         </Route>
+        <Route path="/login" />
+
         <Route exact path="/">
           <FrontPage />
         </Route>
