+++ client/src/components/Login.js
@@ -1,0 +1,1 @@
+import React from 'react';
