+++ client/src/styles/colors.js
@@ -1,9 +1,9 @@
 const Colors = {
-  darkGreen: "#28903A",
-  mediumGreen: "#3fbf7fff",
-  mediumTeal: "#40e3baff",
-  lightBlue: "#64f5e9ff",
-  lightGrey: "#afd2deff",
+  darkGreen: '#28903A',
+  mediumGreen: '#3fbf7fff',
+  mediumTeal: '#40e3baff',
+  lightBlue: '#64f5e9ff',
+  lightGrey: '#afd2deff',
 };
 
 export default Colors;
