+++ client/src/views/FrontPage.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import YourPath from "../components/YourPath";
-import EventsProvider from "../components/Events/EventsProvider";
-import PostsProvider from "../components/Posts/PostsProvider";
+import YourPath from '../components/YourPath';
+import EventsProvider from '../components/Events/EventsProvider';
+import PostsProvider from '../components/Posts/PostsProvider';
 
 export default function FrontPage() {
   return (
@@ -11,5 +11,5 @@ export default function FrontPage() {
       <EventsProvider />
       <PostsProvider />
     </>
-  )
+  );
 }
